Refactored structure with 'Utils' folder
